keyboard and mouse recorder can be played from zeuz
